+++ src/lib.rs
@@ -6,4 +6,6 @@ pub mod particles;
 pub mod prelude {
     pub use crate::audio::*;
     pub use crate::collision::*;
+    pub use crate::colors::*;
+    pub use crate::particles::*;
 }