+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+## [0.2.1] - 2024-08-15
+
 - Add particles and colors items to the prelude
 
 ## [0.2.0] - 2024-08-15
@@ -12,7 +14,8 @@
 - Initial release with `collision_started`, `collision_stopped`, `play_sound`, `particle_trail_bundle`, and `spawn_particle_poof`
 
 <!-- next-url -->
-[Unreleased]: https://github.com/CleanCut/bevy_cleancut/compare/v0.2.0...HEAD
+[Unreleased]: https://github.com/CleanCut/bevy_cleancut/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/CleanCut/bevy_cleancut/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/CleanCut/bevy_cleancut/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/CleanCut/bevy_cleancut/compare/v0.0.0...v0.1.0